+++ components/game/firebase/addToGameCollection.ts
@@ -2,7 +2,7 @@ import { addDoc, collection, doc, setDoc, Timestamp } from 'firebase/firestore'
 import { db } from '../../../firebase/clientApp'
 import generateGameConfig from '../constants/BaseGameConfig'
 import gamesConverter from '../../../firebase/converters/gamesConverter'
-import updateUserGamesList from '../../../firebase/fetch/updateUserGamesList'
+import updateUserGamesList from './updateUserGamesList'
 
 export const generateBoardDataDocRef = (gameID: string, userID: string) =>
   // Path: games (collection) -> gameID (document) -> userID (subcollection) -> boardData (document)
